// frontend/src/app/data-type.ts
// user interface
export interface userProfileInterface {
  name: string;
  email: string;
  contect: string;
}

// card interface
export interface cardInterface {
  tid: String;
  title: string;
  category: string;
  email: string;
  phoneNumber: number;
  address: string;
  description: string;
  linkdinLink: string;
  facebookLink: string;
  instagramLink: string;
  websiteLink: string;
}